Move sansio code to a sansio module
This makes the naming a little easier (called request and response),
and allows the sansio module to be considered private/development. The
latter is desired as it isn't clear (yet) how to specify the IO
interface - in an abstract manner so that both sync and async
implementations exist.

This also allows further sansio, rather than WSGI based code, to be
added in a clear location - clear to future authors that the code must
not be WSGI or ASGI or IO specific.
